Fixed dynamic-refs

+++ bin/cache.fsx
@@ -8,9 +8,7 @@ let args = Args.getArgs() |> Array.toSeq |> Seq.skip 1 |> Seq.toArray
 
 let globalCacheDir = System.Environment.GetFolderPath(System.Environment.SpecialFolder.ApplicationData) + """\npm\node_modules\nfsr"""
 
-match args with
-| [|"clean"|] -> 
+if Args.contains [|"clean"|] then
     for file in Directory.GetFiles(globalCacheDir, "*.cache") do
         printfn "removing %s" file
         File.Delete(file)
-| _ -> ()
